machine/pyproject.toml
[tool.poetry]
name = "machine"
version = "0.0.1"
description = "HTTP API for machine learn economy price predictions and teach"
authors = ["Renan Dias Santos de Santana, Arthur Roberto Weege Pontes e Guilherme Silveira Fernandes da Silva"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.11.0"
flask = "^3.0.0"
flask-cors = "^4.0.0"
numpy = "^2.3.3"
pandas = "^2.0.0"
scikit-learn = "^1.3.0"
xgboost = "^2.0.0"
lightgbm = "^4.0.0"
yfinance = "^0.2.0"
joblib = "^1.3.0"
python-dotenv = "^1.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start-server = "app.main:main"